ACoRD: Update readme.md - Integrations, Python
Github/IrisDroidology/Python-Learning: right now, with Impractical Python Projects (IPyP), we're doing #14 - marsorbiter.py. We will update this game so that there are more planets, including Midgard, which is a metallic coloured planet that shines with different coloured light (reflective) depending on which parts are pointed to the sun. This allows massive energy stores to be collected. Star Sailors - acord.tech/readstarsailors
Belletristica